@@ -24,6 +24,7 @@
     "symfony/flex": "^1.0",
     "symfony/form": "^4.1",
     "symfony/framework-bundle": "^4.1",
+    "symfony/mailer": "^4.1",
     "symfony/maker-bundle": "^1.4",
     "symfony/monolog-bundle": "^3.1",
     "symfony/orm-pack": "^1.0",
@@ -31,7 +32,6 @@
     "symfony/requirements-checker": "^1.0",
     "symfony/security-bundle": "^4.1",
     "symfony/serializer": "^4.1",
-    "symfony/swiftmailer-bundle": "^3.1",
     "symfony/translation": "^4.1",
     "symfony/twig-bundle": "^4.1",
     "symfony/validator": "^4.1",
